#5984. Minor code cleanup.
1) Store multiloot_lastfrob as int, since gameLocal.time is saved into it, which is int too.
2) Simplfy multuloot + noinv items: check for inv. item once.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9944 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/Player.h
@@ -859,7 +859,7 @@ public:
 
 	// Obstorte: #5984 (multilooting)
 	bool					multiloot;
-	float					multiloot_lastfrob;
+	int						multiloot_lastfrob;
 
 	// angua: Set ideal crouch state
 	void					EvaluateCrouch();
